Refactored package and main activity names
