+++ android/build.gradle
@@ -15,7 +15,7 @@ android {
         instrumentTest.setRoot('tests')
     }
     defaultConfig {
-        applicationId "com.icarus.prototype"
+        applicationId "com.icarus.project"
         minSdkVersion 8
         targetSdkVersion 24
     }
@@ -68,7 +68,7 @@ task run(type: Exec) {
     }
 
     def adb = path + "/platform-tools/adb"
-    commandLine "$adb", 'shell', 'am', 'start', '-n', 'com.icarus.prototype/com.icarus.prototype.AndroidLauncher'
+    commandLine "$adb", 'shell', 'am', 'start', '-n', 'com.icarus.project/com.icarus.project.AndroidLauncher'
 }
 
 // sets up the Android Eclipse project, using the old Ant based build.
