+++ android/src/com/icarus/project/AndroidLauncher.java
@@ -1,16 +1,15 @@
-package com.icarus.prototype;
+package com.icarus.project;
 
 import android.os.Bundle;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
-import com.icarus.prototype.IcarusPrototype;
 
 public class AndroidLauncher extends AndroidApplication {
 	@Override
 	protected void onCreate (Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
-		initialize(new IcarusPrototype(), config);
+		initialize(new ProjectIcarus(), config);
 	}
 }
