+++ core/src/com/icarus/project/Waypoint.java
@@ -1,11 +1,10 @@
-package com.icarus.prototype;
+package com.icarus.project;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.utils.Align;
 import com.google.gson.JsonObject;
-import com.icarus.prototype.Colors;
 
 class Waypoint {
     public String name;
