+++ ios/build.gradle
@@ -4,7 +4,7 @@ sourceCompatibility = '1.7'
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
 ext {
-	mainClassName = "com.icarus.prototype.IOSLauncher"
+	mainClassName = "com.icarus.project.IOSLauncher"
 }
 
 launchIPhoneSimulator.dependsOn build
