+++ ios/robovm.properties
@@ -1,6 +1,6 @@
 app.version=1.0
 app.id=com.icarus.prototype
-app.mainclass=com.icarus.prototype.IOSLauncher
+app.mainclass=com.icarus.project.IOSLauncher
 app.executable=IOSLauncher
 app.build=1
 app.name=icarus-prototype
